Simplify component moving functions to utils
